+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - lots of work on bubbling up errors.  InfraErrors treated differently.  Standardised all errors under ops/
 - split the Ic ledger code over api, workflow, ops and infra
 - re-wrapped the ic Call type via CallOps, adding metrics
+- moved IC management status view adapters into workflow; ops now return internal mgmt status types
 
 ## [0.7.12] - 2026-01-05
 - updated the signature code to use the HashTree from ic-certified-map.  Basically none of us know how it works so this